+++ Frames/PaddingPlaceholderFrame.cs
@@ -1,11 +1,11 @@
 ﻿namespace Mp3TagReader.Frames {
 	internal class PaddingPlaceholderFrame : IFrame {
 
-		public PaddingPlaceholderFrame( ulong size ) {
+		public PaddingPlaceholderFrame( int size ) {
 			Size = size;
 		}
 
 		public string Id => "0000 (Padding placeholder)";
-		public ulong Size { get; }
+		public int Size { get; }
 	}
 }
